feat: moved observation files
